// src/tasks/BeanstalkCreateApplicationVersion/types.gen.ts
/*!
 * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * SPDX-License-Identifier: MIT
 */

    
/* NOTE! this file is auto-generated by generateResources.ts, do not edit it by hand. */


/*!
 * Copyright 2019 Amazon.com, Inc. or its affiliates. All Rights Reserved.
 * SPDX-License-Identifier: MIT
 */

    
/* NOTE! this file is auto-generated by generateResources.ts, do not edit it by hand. */

export type BeanstalkCreateApplicationVersionInput = {
    awsCredentials: string
    regionName: string
    applicationName: string
    applicationType: 'aspnet' | 'aspnetCoreWindows' | 'aspnetCoreLinux' | 's3'
    webDeploymentArchive: string
    dotnetPublishPath: string
    deploymentBundleBucket: string
    deploymentBundleKey: string
    versionLabel: string
    description: string
    outputVariable: string
    logRequest: boolean
    logResponse: boolean
}

export type BeanstalkCreateApplicationVersionMessages = 
    | 'DisplayApplicationType'
    | 'CreatedApplicationVersion'
    | 'CreatedApplicationVersionWithDescription'
    | 'DeterminedBucket'
    | 'UploadingBundle'
    | 'BundleUploadCompleted'
    | 'BundleUploadFailed'
    | 'WaitingForDeployment'
    | 'FailedToFindEnvironment'
    | 'CreatingBeanstalkBundle'
    | 'UsingExistingBeanstalkBundle'
    | 'ArchiveSize'
    | 'ZipError'
    | 'PublishingPath'
    | 'AddingAspNetCoreBundle'
    | 'AddingFilesFromDotnetPublish'
    | 'AddingManifest'
    | 'BundleComplete'
    | 'ApplicationDoesNotExist'
    | 'SettingOutputVariable'
    | 'ApplicationExistsQueryError'
    | 'ApplicationExistsQueryErrorTooManyApplications'
    | 'TaskCompleted'